apps/streamlit/requirements.txt
streamlit
pandas
Pillow
opencv-python-headless
pytesseract
numpy
matplotlib
